--- src/appp.jsx
@@ -1,34 +1,0 @@
-// import React from "react";
-// import "./App.css";
-// import Header from "./Components/Header";
-
-// function App() {
-//   return (
-//     <div className="App">
-//       <div>
-//         <p></p>
-//       </div>
-//     </div>
-//   );
-// }
-
-// export default App;
-
-// import React from "react";
-// import Comments from "./component/forumcomment/Comments";
-// import { BrowserRouter, Route } from "react-router-dom";
-// import home from "./component/POTD/home";
-// import nasaphoto from "./component/POTD/nasaphoto";
-
-// function App() {
-//   return (
-//     <BrowserRouter>
-//       <div>
-//         <Route component={home} path="/" exact />
-//         <Route component={nasaphoto} path="/nasaphoto" />
-//       </div>
-//     </BrowserRouter>
-//   );
-// }
-
-// export default App;
