+++ src/component/POTD/potd.css
@@ -1,0 +1,30 @@
+/* body {
+  margin: 0;
+  font-family: "Times New Roman", "Times", "serif";
+  background-position: center;
+  background-size: cover;
+  background-repeat: no-repeat;
+  color: rgb(248, 243, 243);
+  display: flex;
+  justify-content: center;
+}
+
+div {
+  padding-top: 50px;
+  padding-right: 80px;
+  padding-bottom: 50px;
+  padding-left: 80px;
+} */
+
+.Title {
+  color: rgb(250, 248, 245);
+  font-weight: bold;
+}
+.tarik {
+  color: rgb(6, 207, 252);
+  font-family: "Lucida Console", "Courier New", monospace;
+}
+
+.description {
+  color: rgb(255, 255, 255);
+}
