+++ src/component/forumcomment/CommentForm.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useState } from "react";
+import "./index.css";
 
 const CommentForm = ({
   handleSubmit,
