+++ src/component/forumcomment/index.css
@@ -1,0 +1,4 @@
+.comments-title {
+  color: rgb(205, 123, 0);
+  font-weight: bold;
+}
